Updating database schema.
I am updating the database schema a bit. For one, we don't need a hundred files to initialize a database with only a few tables, and for another,
I am normalizing the naming of things a bit. Next I will be deleting the excess .sql files, and updating the currently existing classes to represent 
the new schema. It is better to do this now when it ultimately only impacts one class, than when the project is almost complete.