Updated school.go
In theory, we now have the ability to create and retrieve Schools as well as 
the ability to retrieve a list of spells associated with the School.  This all
needs testing before I actually decide it works. The School "type" is also
now capable of returning a link to itself, assuming it is properly formed.